chore: Make dev and prod servers use same UID
There's no script to stop the dev server but there is one to stop the
prod server.

Since it's highly unlikely that the dev and prod servers will be run at
the same time, make them use the same UID so that they can both be
stopped by the same script.