fix(profiles): Convert handles to hex
tc requires handles to be hexadecimal rather than decimal. This
requirement isn't documented anywhere, however...

Not converting handles to hexadecimal causes the setup script to fail
once there are seven or more profiles: the root prio qdisc has the
correct number of classes (10) but the seventh child qdisc refers to the
16th (0x10), which doesn't exist.

See #13 for context.